requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   cattrs
    #   requests-cache
cattrs==23.2.3
    # via requests-cache
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
huggingface-hub==0.23.4
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   fastapi
    #   torch
joblib==1.4.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
orjson==3.10.6
    # via fastapi
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pillow==10.4.0
    # via sentence-transformers
platformdirs==4.2.2
    # via requests-cache
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.5.15
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   requests-cache
    #   tiktoken
    #   transformers
requests-cache==1.2.1
    # via -r requirements.in
rich==13.7.1
    # via typer
safetensors==0.4.3
    # via transformers
scikit-learn==1.5.1
    # via sentence-transformers
scipy==1.14.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via url-normalize
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
sympy==1.12.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.7.0
    # via -r requirements.in
tokenizers==0.19.1
    # via transformers
torch==2.3.1
    # via sentence-transformers
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.42.3
    # via sentence-transformers
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
ujson==5.10.0
    # via fastapi
url-normalize==1.4.3
    # via requests-cache
urllib3==2.2.2
    # via
    #   requests
    #   requests-cache
uvicorn==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
